+++ src/ProductConstructionService/ProductConstructionService.Deployment/Utility.cs
@@ -1,6 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.IO.Compression;
+using System.Web;
+
 namespace ProductConstructionService.Deployment;
 
 internal static class Utility
@@ -11,4 +14,17 @@ internal static class Utility
         return true;
     }
 
+    public static string ConvertStringToCompressedBase64EncodedQuery(string query)
+    {
+        var bytes = System.Text.Encoding.UTF8.GetBytes(query);
+        MemoryStream memoryStream = new();
+        GZipStream compressedStream = new(memoryStream, CompressionMode.Compress);
+
+        compressedStream.Write(bytes, 0, bytes.Length);
+        compressedStream.Close();
+        memoryStream.Seek(0, SeekOrigin.Begin);
+        var data = memoryStream.ToArray();
+        var base64query = Convert.ToBase64String(data);
+        return HttpUtility.UrlEncode(base64query);
+    }
 }
